:recycle: refactor project layout
