+++ pyproject.toml
@@ -67,11 +67,6 @@ preview = true
 select = ["ALL"]
 ignore = [
     "CPY",     # Missing copyright header
-    "D100",    # Missing docstring in public module
-    "D101",    # Missing docstring in public class
-    "D102",    # Missing docstring in public method
-    "D103",    # Missing docstring in public function
-    "D104",    # Missing docstring in public package
     "D107",    # Missing docstring in __init__
     "D203",    # 1 blank line required before class docstring
     "D211",    # No blank lines allowed before class docstring
