+++ rh_aws_saml_login/__init__.py
@@ -1,0 +1,5 @@
+"""Expose the public API of the package."""
+
+from ._models import AwsCredentials
+
+__all__ = ["AwsCredentials"]
