+++ rh_aws_saml_login/_consts.py
